Pushed previews into separate functional component and fixed the dialog formatting
